+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ### 1.0.0: Aug 3rd, 2015
-*
-*
+* Added `add_theme_support('rankz-init');` for first Shaba setup
+* Added `add_theme_support('rankz-clean-up');`, a bunch of functions to clean up WordPress dashboard
+* Added `add_theme_support('rankz-disable-comments');` to disable comments on whole website
+* Added `add_theme_support('rankz-disable-widgets');` to disable widgets on whole website
+* Added `add_theme_support('rankz-disable-editors');` to disable editors on WordPress dashboard
+* Added `add_theme_support('rankz-remove-default-image-sizes');` to remove default WordPress image sizes and hide Media page in WordPress settings
+* Added `add_theme_support('rankz-menu-humility');`, a snippet to push all Custom Post Types at the bottom of the admin menu
+* Added `add_theme_support('rankz-font-awesome-menu');` to add Font Awesome icons in the WordPress menus
+* Added `add_theme_support('rankz-admin-login', 'URL', '#HEX-COLOR');` to customize WordPress admin login page with differents colours, custom logo and custom url link for logo
+* Added `add_theme_support('rankz-bootstrap-nav-walker');` an alternative to soil-nav-walker with bootstrap structures and classes